app/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
redis
python-dotenv
jinja2